src/main/webapp/app/entities/eRestaurantCustomer/on-hold/route/on-hold-routing-resolve.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, ActivatedRoute, Router, convertToParamMap } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { IOnHold, OnHold } from '../on-hold.model';
import { OnHoldService } from '../service/on-hold.service';

import { OnHoldRoutingResolveService } from './on-hold-routing-resolve.service';

describe('OnHold routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: OnHoldRoutingResolveService;
  let service: OnHoldService;
  let resultOnHold: IOnHold | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            snapshot: {
              paramMap: convertToParamMap({}),
            },
          },
        },
      ],
    });
    mockRouter = TestBed.inject(Router);
    jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
    routingResolveService = TestBed.inject(OnHoldRoutingResolveService);
    service = TestBed.inject(OnHoldService);
    resultOnHold = undefined;
  });

  describe('resolve', () => {
    it('should return IOnHold returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 'ABC' };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultOnHold = result;
      });

      // THEN
      expect(service.find).toBeCalledWith('ABC');
      expect(resultOnHold).toEqual({ id: 'ABC' });
    });

    it('should return new IOnHold if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultOnHold = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultOnHold).toEqual(new OnHold());
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as OnHold })));
      mockActivatedRouteSnapshot.params = { id: 'ABC' };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultOnHold = result;
      });

      // THEN
      expect(service.find).toBeCalledWith('ABC');
      expect(resultOnHold).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
